requirements.txt
numpy~=1.25.2
openpyxl~=3.1.2
pandas~=2.0.3
python_dateutil
Twisted==22.10.0
Scrapy==2.10.0
pathvalidate~=3.1.0
PyPDF2
name_matching
PyMuPDF